jewels-jewellery-ecommerce-free-website-template/requirements.txt
Flask==3.1.0
gunicorn==23.0.0
mysql-connector-python==9.2.0
